Finishing cabins CRUD and App settings

+++ src/services/apiCabins.js
@@ -11,20 +11,32 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  
-  //https://pxenfuasnkmuieedgdvz.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+//https://pxenfuasnkmuieedgdvz.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
   // 1. create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  let query = supabase.from("cabins");
+
+  // a. Create
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // b. edit
+  if (id)
+    query = query
+      .update({ ...newCabin, image: imagePath })
+      .eq("id", id)
+      .select();
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
@@ -32,15 +44,19 @@ export async function createCabin(newCabin) {
   }
 
   // 2. upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
   // 3. Delete cabin if there was an error uploading the image
-  if(storageError) {
+  if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
-    throw new Error("Cabin image failed to upload hence cabin didnt get created😭");
+    throw new Error(
+      "Cabin image failed to upload hence cabin didnt get created😭"
+    );
   }
 
   return data;
